On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	new file:   .gitignore

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	new file:   scripts/main.ps1

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   .devcontainer/devcontainer.json

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Author:    waraken <a_m_@hotmail.co.jp>
Date:      Tue Aug 12 12:06:09 2025 +0900

interactive rebase in progress; onto 25744ea
Last commands done (3 commands done):
   squash 6479aad On branch main Your branch is ahead of 'origin/main' by 1 commit.   (use "git push" to publish your local commits)
   squash a1237fe On branch main Your branch is up to date with 'origin/main'.
No commands remaining.
You are currently rebasing branch 'waraken' on '25744ea'.

Changes to be committed:
	new file:   .devcontainer/devcontainer.json
	new file:   .gitignore
	new file:   scripts/main.ps1