On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	new file:   .vscode/launch.json
	new file:   resource/only-original.txt
	new file:   scripts/diff.ps1
	new file:   scripts/download_two_tag_file.ps1
	new file:   scripts/format.ps1
	deleted:    scripts/main.ps1
	new file:   temp/diff.txt

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	renamed:    resource/only-original.txt -> resource/only-original.csv

On branch main
Your branch is ahead of 'origin/main' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   autocomplete.txt
	modified:   scripts/format.ps1

On branch main
Your branch is ahead of 'origin/main' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   LICENSE

On branch main
Your branch is ahead of 'origin/main' by 4 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   README.md

On branch main
Your branch is ahead of 'origin/main' by 5 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   scripts/download_two_tag_file.ps1

On branch main
Your branch is ahead of 'origin/main' by 6 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	renamed:    resource/only-original.csv -> resource/only-contains-comfyui-custom-scripts.csv

On branch main
Your branch is ahead of 'origin/main' by 7 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   scripts/format.ps1

On branch main
Your branch is ahead of 'origin/main' by 8 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   scripts/diff.ps1

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   scripts/make-lite.ps1

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   autocomplete-lite.txt

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   scripts/make-lite.ps1

On branch main
Your branch is ahead of 'origin/main' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   README.md

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Tue Aug 12 23:43:32 2025 +0000

interactive rebase in progress; onto de3ab21
Last commands done (13 commands done):
   squash a090754 On branch main Your branch is ahead of 'origin/main' by 1 commit.   (use "git push" to publish your local commits)
   squash 51f6aef On branch main Your branch is ahead of 'origin/main' by 2 commits.   (use "git push" to publish your local commits)
No commands remaining.
You are currently rebasing branch 'waraken' on 'de3ab21'.

Changes to be committed:
	new file:   .vscode/launch.json
	modified:   LICENSE
	modified:   README.md
	new file:   autocomplete-lite.txt
	modified:   autocomplete.txt
	new file:   resource/only-contains-comfyui-custom-scripts.csv
	new file:   scripts/diff.ps1
	new file:   scripts/download_two_tag_file.ps1
	new file:   scripts/format.ps1
	deleted:    scripts/main.ps1
	new file:   scripts/make-lite.ps1
	new file:   temp/diff.txt

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	new file:   scripts/main.ps1

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   .vscode/launch.json
	new file:   resource/only-original.txt
	new file:   scripts/diff.ps1
	new file:   scripts/download_two_tag_file.ps1
	new file:   scripts/format.ps1
	deleted:    scripts/main.ps1
	new file:   temp/diff.txt

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	renamed:    resource/only-original.txt -> resource/only-original.csv

On branch main
Your branch is ahead of 'origin/main' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   LICENSE

On branch main
Your branch is ahead of 'origin/main' by 4 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   README.md

On branch main
Your branch is ahead of 'origin/main' by 5 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   scripts/download_two_tag_file.ps1

On branch main
Your branch is ahead of 'origin/main' by 6 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	renamed:    resource/only-original.csv -> resource/only-contains-comfyui-custom-scripts.csv

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   scripts/make-lite.ps1

On branch waraken
Your branch is ahead of 'origin/waraken' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	deleted:    resource/only-original.txt
	modified:   scripts/diff.ps1